Use server schema to improve DependencyOrder
This change uses server-provided schema information to improve the
dependency-order heuristic in several important ways:

- Special-case CRD/TPR objects and sort them first.
- Sorts all non-namespaced resources before namespaced.
- Finds all (namespaced) types that mention v1.PodSpec and sorts them
  last.

This approach is generic and covers significantly more types than the
previous logic.  The only special case is now CRD/TPRs.

The downside is that the schema fetch is slow, so this change adds an
Info message to let the user know why they are waiting.