Embed kubecfg.libsonnet inside the kubecfg binary
This change adds a new "internal:///" URL to the end of the search
path, based on files embedded in the binary using go-bindata.
Currently this only includes `kubecfg.libsonnet`.

Note the new generated `utils/bindata.go` is added to git so a naive
"go get ..." is still possible.  Run `make generate` to regenerate
this file when needed (and include any changes in PR).