Support multiple `-J` args
Previously only a single colon-separated `-J` arg was accepted, and
treated like `KUBECFG_JPATH`.  This was a bit weird and hard to use.

This change allows `-J` to be given multiple times, each time being a
simple complete path (no internal separator) - as it should always
have been.  For similar reasons, `-U` is also changed from a
`StringSlice` to a `StringArray`.