Validate input before `update`
Validate the input objects against the server-side schema before
starting an update.  This is the exact equivalent of a previous
`kubecfg validate $input && kubecfg update $input`

In particular, this will reject unrecognised JSON keys that would
previously be silently ignored by the apiserver.  It will also catch a
few classes of error (eg: wrong value types) before starting the
update - previously these would be rejected by the server, aborting
the update midway.

I'm pretty sure everyone thought something like this was already
happening, so hopefully it's a no-op.  If it hurts anyone, there's a
new `--validate=false` flag to go back to the old behaviour.

(This PR is possible thanks to the earlier validate improvements in #199)