Re-add --ignore-unknown flag to validate
Sometimes we want to validate a config that contains CRDs that don't
yet exist on the server, or for whatever reason perform only a "best
effort" validation against the server's time-shifted idea of the
schema.

There are several ways we could attempt to anticipate the future
creation of such CRDs (eg: by inspecting the set of objects we are
about to create), but there is always going to be corner case that is
not covered by such logic.

This change just re-adds a `validate --ignore-unknown`
flag (originally from #191), which allows validate to skip unknown
kinds and proceed with remaining known kinds.

See also original issue #146