Properly detect equal diffs and change regex definition to be static.
This addresses code review comments -- the conditions in original PR
were detected incorrectly; this PR changes the condition to the case
where the whole Diff is just a single block of DiffEquals change.

Also, the latter change moves DiffLineStart to the top of file,
so that it doesn't need to be recomputed each time the formatDiff
function is called.

Signed-off-by: Milan Plzik <milan.plzik@ceai.io>