Switch to strategic-merge-patch
If an object does not yet exist on the server, kubecfg creates it
using a `POST` - this has not changed.

If an object already exists on the server, previously kubecfg would
`PATCH` it with a JSON merge patch to produce the updated version.
This was simple and worked for _almost all_ cases, but (eg) failed to
remove annotations that had been removed from local config.  This
latter is surprising because it is not how `kubectl apply` works, and
it is not the object-level semantics that `kubecfg` implements with
its own garbage collection.

This change switches to performing a schema-driven client-side 3-way
strategic-merge-patch between the last version of the object created
by `kubecfg`, the new version of the object found in local configs,
and the current version found on the server.  The "last version" of
the object is stored JSON-serialised in a new
`kubecfg.ksonnet.io/last-applied-configuration` annotation on each
object.  The updated object is then sent to the server using a
`PUT` (with retry on resource conflict).

If schema information is not available (CRDs), then it will fallback
to json-merge (as before this change, but client-side).

This reimplements the `kubectl apply` semantics, and is consistent with
the other kubernetes tools in this space.

Also a few minor but user-visible verbosity improvements:
- Now prints only one of "Creating ..." or "Updating ..." progress
  message.
- Prints no message at all if the object has not changed (according to
  `Semantic.DeepEqual`)

Fixes #226