Use more compact encoding for last-applied-configuration
This annotation contains an approximate copy of the object itself, and
can be just as large (or even larger when including JSON vs protobuf
encoding overhead).  This effectively *halves* the maximum k8s object
size when using strategic-merge-patch, which is a big deal for things
like ConfigMaps, and generally wastes network traffic on every API
read/write round trip.

Note the `kubecfg.ksonnet.io/last-applied-configuration` annotation is
only used by `kubecfg`, so there is no need to be compatible with the
similar annotation used by `kubectl apply`.

This change switches to a smaller encoding: base64 encoded, gzipped,
compact JSON.  The downside is that it is harder for humans to
manually inspect the encoded data.

For reference, the "manual" way to inspect this attribute is:
```
kubectl get deployment/foo -o jsonpath='{.metadata.annotations.kubecfg\.ksonnet\.io/last-applied-configuration}' |
  base64 -d | zcat | jq
```

Thanks to @seh for the suggestion!