Try JSON merge when strategic merge prep. fails
As of Kubernetes version 1.15.2, the API server can't accept strategic
merge patches for custom resources. Since it is difficult to discern
which objects belong to custom resources, instead interpret a failure
to prepare a strategic merge patch as a likely indicator that the
API server does not support the patch type for the object, and fall
back to trying a JSON merge patch instead.