Fix handling of schema-less CRD objects in k8s >=1.15
The server-side handling of CRD schemas changed in Kubernetes 1.15.  A
schema-less CRD is now published in the openapi schema as an empty map
of `type:object`, where previously it didn't appear at all.

This trivial schema confused the strategicmergepatch code, leading to
an `expected kind, but got map` error, because there are no additional
properties defined.

Update the client-side patch code to detect these "empty" schemas by
testing for the presence of a `metadata` property.  Now an empty
schema is treated as if it was not found, and we fallback to the old
just-replace logic.

Fixes #259