Detect imported extvars and TLAs via sentinel path
When a Jsonnet consults an Importer for a path nominated by an
external variable or top-level argument file-based command-line flag,
the "importedFrom" parameter passed to the Importer.Import method is
not a valid URL, but is instead a sentinel placeholder value (namely,
an empty string). Detect these sentinel values and use the current
working directory as the base URL instead when resolving relative
import paths.