add a native function to create an NT(LM) hash for a password.
The 'ntHashFromPassword' native function is added to kubecfg.libsonnet.
It can be used to calculate the NT(LM) hash for a password.
This is useful when these kinds of hashes need to be created in order
to add them to some initial configurations for applications that
expect this format (e.g. Alfresco DMS).