add an example library to integrate kubecfg with 1Password
This commit includes '1password.libsonnet' that builds on the
new features commited before to kubecfg like:
- generatePassword
- execProgram
- ntHashFromPassword
- encodeBase64Url

This library can be used to generate password for secrets.
These password are then stored in a vault in 1Password.
Once the new item containing the secrets is stored in the
1Password vault kubecfg will use these when 'rendering'
the secret resource.

In addition to the library 'secrets.jsonnet' is provided as an
example on how to use the library.