diff: Add --no-error-on-diff flag
Add a `--no-error-on-diff` flag to the `diff` subcommand so as to not
exit with an error when there are diffs present. It will still exit with
an error if the configs are not valid, so `kubecfg diff` can be a useful
command to run on a branch of a CI system to show diffs and still
validate the configs.